+++ R/YamlContext.ReactiveReference.S3.r
@@ -1,5 +1,5 @@
 #' @title
-#' Class: ReactiveReferenceYamlContext.S3 
+#' Class: YamlContext.ReactiveReference.S3 
 #'
 #' @description
 #' Class representing the reactive reference context with respect to 
@@ -9,28 +9,28 @@
 #'
 #' @param .x \code{\link{ANY}}. An object of an arbitrary class whose class
 #'    attribute should be updated so that it becomes an instance of class
-#'    \code{ReactiveReferenceYamlContext.S3}. Mainly intended for rapid prototyping 
+#'    \code{YamlContext.ReactiveReference.S3}. Mainly intended for rapid prototyping 
 #'    purposes
 #'    
 #' @field ctx \code{\link{character}}.
 #'    Context ID. Corresponds to the class name.
-#'    Initial: \code{"ReactiveReferenceYamlContext.S3"}.
-#' @return Instance of class \code{ReactiveReferenceYamlContext.S3}.
-#' @example inst/examples/ReactiveReferenceYamlContext.S3.r
+#'    Initial: \code{"YamlContext.ReactiveReference.S3"}.
+#' @return Instance of class \code{YamlContext.ReactiveReference.S3}.
+#' @example inst/examples/YamlContext.ReactiveReference.S3.r
 #' @template author
 #' @template references
 #' @export
-ReactiveReferenceYamlContext.S3 <- function(
+YamlContext.ReactiveReference.S3 <- function(
   .x,
-  ctx = "ReactiveReferenceYamlContext.S3"
+  ctx = "YamlContext.ReactiveReference.S3"
 ) {
   if (!missing(.x)) {
-    class(.x) <- c("ReactiveReferenceYamlContext.S3", class(.x))
+    class(.x) <- c("YamlContext.ReactiveReference.S3", class(.x))
     out <- .x
   } else {
     out <- new.env()
     out$ctx <- ctx
-    class(out) <- c("ReactiveReferenceYamlContext.S3", class(out))
+    class(out) <- c("YamlContext.ReactiveReference.S3", class(out))
   }
   return(out)
 }