+++ inst/examples/YamlContext.ReactiveReference.S3.r
@@ -2,24 +2,24 @@
 
 ## Informal use (intended mainly for rapid prototyping) //
 ## Takes *any* object and simply changes the class attributes
-ReactiveReferenceYamlContext.S3(
+YamlContext.ReactiveReference.S3(
   list(
     ctx = "abcd"
   )
 )  
-ReactiveReferenceYamlContext.S3(TRUE)  
+YamlContext.ReactiveReference.S3(TRUE)  
 
 ## Formal use (explicitly using 'fields') //
-res <- ReactiveReferenceYamlContext.S3()
+res <- YamlContext.ReactiveReference.S3()
 ls(res)
 res$ctx
-res <- ReactiveReferenceYamlContext.S3(
-  ctx = "ReactiveReferenceYamlContext.S3"
+res <- YamlContext.ReactiveReference.S3(
+  ctx = "YamlContext.ReactiveReference.S3"
 )
 res$ctx
 
 ## Recommended: include namespace //
 ## Regardless if you plan on using this class in an informal or formal way
-fromyaml::ReactiveReferenceYamlContext.S3()
+fromyaml::YamlContext.ReactiveReference.S3()
 
 }