+++ tests/testthat/test-parseYaml.r
@@ -5,7 +5,7 @@ test_that("parseYaml", {
     from = function() {
     "reactive-ref: {id: x_1, where: .GlobalEnv, as: ref_1}"
     },
-    ctx = ReactiveReferenceYamlContext.S3()
+    ctx = YamlContext.ReactiveReference.S3()
   )
   yaml
   expect_is(res <- parseYaml(yaml = yaml), "ReactiveReferenceYamlParsed.S3")