+++ NEWS.md
@@ -1,3 +1,22 @@
+# CHANGES IN fromyaml VERSION 0.3.2
+
+## NEW FEATURES
+
+## BUG FIXES
+
+## MAJOR CHANGES
+
+## MINOR CHANGES
+
+- added: `where` in `getYaml()`
+- Renamed: `ReactiveReference.S3` --> `YamlContext.ReactiveReference.S3` 
+- modified: `README.md` 
+- modified: examples
+
+## MISC
+
+-----
+
 # CHANGES IN fromyaml VERSION 0.3.1
 
 ## NEW FEATURES